Fix OTLP 2 STEF converter
Add regen-testdata makefile target that uses the tool
and makes recreating STEF files after format changes easier.

Had to fix hipstershop-otelmetrics.zst file to prepend the
file size to it so that the tool can work with it. Now all
test files use size-prefix format.
